refactor: naming styles files
