+++ packages/library/components/detail/src/detail-styles.css
