+++ packages/library/components/icon/src/icon-styles.css
