+++ packages/library/components/inputter/src/inputter-styles-detail.css
