+++ packages/library/components/inputter/src/inputter-styles.slotted.css
@@ -2,7 +2,7 @@
   /* NOTE: targeting Safari only */
   @media not all and (min-resolution: 0.001dpcm) {
     /*
-      NOTE: These styles are duplicated from inputter/styles.css
+      NOTE: These styles are duplicated from inputter-styles.css
       due to this bug in Webkit: https://bugs.webkit.org/show_bug.cgi?id=223814
     */
     & input[type="radio"] {
